requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
joblib==1.2.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
threadpoolctl==3.1.0
torch==1.13.1+cu117
torchvision==0.14.1+cu117
typing_extensions==4.5.0
urllib3==1.26.15
